流的重构和优化，F5和F6优化

+++ WinFormsApp1/Service/IServiceF6.cs
@@ -5,5 +5,5 @@ namespace TaxiManager.Service;
 public interface IServiceF6
 {
     public static IServiceF6 Instance => ServiceF6.Instance;
-    public int GetFlowTo(PositionRange toRange, DateTime time);
+    public float GetFlowTo(PositionRange toRange, DateTime time);
 }
